Changed all node tests to fit refactoring
