+++ src/components/Nodes/LearningPathLearningElementNode/LearningPathLearningElementNode.ts
