+++ src/components/Nodes/NodeTypes/NodeTypes.ts
@@ -10,7 +10,7 @@ import SummaryNode from '../SummaryNode/SummaryNode'
 import VideoNode from '../VideoNode/VideoNode'
 import BasicNode from '../BasicNode/BasicNode'
 import { NodeTypes } from 'reactflow'
-import type { LearningPathLearningElementNode } from './LearningPathLearningElementNode/LearningPathLearningElementNode'
+import type { LearningPathLearningElementNode } from '../LearningPathLearningElementNode/LearningPathLearningElementNode'
 import ApplicationExampleNode from '../ApplicationExampleNode/ApplicationExampleNode'
 import ForumNode from '../ForumNode/ForumNode'
 import LearningObjectiveNode from '../LearningObjectiveNode/LearningObjectiveNode'
@@ -39,5 +39,3 @@ export const nodeTypes: NodeTypes = {
   GROUP: DefaultGroup,
   DEFAULT: BasicNode
 }
-
-export type { LearningPathLearningElementNode }
