+++ src/components/Nodes/index.ts
@@ -1,2 +1,2 @@
 export { nodeTypes } from './NodeTypes/NodeTypes'
-export type { LearningPathLearningElementNode } from './NodeTypes/NodeTypes'
+export type { LearningPathLearningElementNode } from './LearningPathLearningElementNode/LearningPathLearningElementNode'
